🔨 Refactor common module to add map_indicator_path_to_id function
