+++ etl/collections/common.py
@@ -1,9 +1,12 @@
 """Common tooling for MDIMs/Explorers."""
 
 from sqlalchemy.engine import Engine
+from sqlalchemy.orm import Session
 
+import etl.grapher.model as gm
 from etl.collections.model import Collection
 from etl.collections.utils import validate_indicators_in_db
+from etl.db import get_engine
 
 
 def validate_collection_config(collection: Collection, engine: Engine, tolerate_extra_indicators: bool) -> None:
@@ -24,3 +27,16 @@ def validate_collection_config(collection: Collection, engine: Engine, tolerate_
     # Check that all indicators in mdim exist
     indicators = collection.indicators_in_use(tolerate_extra_indicators)
     validate_indicators_in_db(indicators, engine)
+
+
+def map_indicator_path_to_id(catalog_path: str) -> str | int:
+    # Check if given path is actually an ID
+    if str(catalog_path).isdigit():
+        return catalog_path
+
+    # Get ID, assuming given path is a catalog path
+    engine = get_engine()
+    with Session(engine) as session:
+        db_indicator = gm.Variable.from_id_or_path(session, catalog_path)
+        assert db_indicator.id is not None
+        return db_indicator.id
